Create README.md
Here is a repository for all submissions to the Elmhurst College Jay Jam 2017.

The names of the folders are the same as the name of the people who submitted it to me, but keep in mind that there was many games created by teams.

You can download either all projects or download individual projects. 

Some of the projects might have to be built/ran with a game engine.